Modern-SQL/queries/requirements.txt
python-dotenv # attention à ce nom
psycopg
psycopg_pool
tqdm
